Move FaaS related methods into new FaasController
As described by https://github.com/UST-MICO/mico/issues/762 we want
to split up the message receiving and the FaaS logic.
