+++ src/main/java/io/github/ust/mico/kafkafaasconnector/MessageListener.java
@@ -19,31 +19,16 @@
 
 package io.github.ust.mico.kafkafaasconnector;
 
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
-import io.cloudevents.json.Json;
 import io.github.ust.mico.kafkafaasconnector.configuration.KafkaConfig;
-import io.github.ust.mico.kafkafaasconnector.configuration.OpenFaaSConfig;
-import io.github.ust.mico.kafkafaasconnector.exception.BatchMicoCloudEventException;
 import io.github.ust.mico.kafkafaasconnector.exception.MicoCloudEventException;
 import io.github.ust.mico.kafkafaasconnector.kafka.MicoCloudEventImpl;
-import io.github.ust.mico.kafkafaasconnector.kafka.RouteHistory;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.client.HttpStatusCodeException;
-import org.springframework.web.client.RestTemplate;
 
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 
 @Slf4j
@@ -51,19 +36,14 @@ import java.util.List;
 public class MessageListener {
 
     @Autowired
-    private RestTemplate restTemplate;
-
-    @Autowired
     private KafkaConfig kafkaConfig;
-
-    @Autowired
-    private OpenFaaSConfig openFaaSConfig;
 
     @Autowired
     private KafkaMessageSender kafkaMessageSender;
 
     @Autowired
-    private CloudEventManipulator cloudEventManipulator;
+    private FaasController faasController;
+
 
     /**
      * Entry point for incoming messages from kafka.
@@ -85,12 +65,8 @@ public class MessageListener {
             if (cloudEvent.getExpiryDate().map(exp -> exp.compareTo(ZonedDateTime.now()) < 0).orElse(false)) {
                 log.debug("Received expired message!");
                 throw new MicoCloudEventException("CloudEvent has already expired!", cloudEvent);
-            } else if (this.openFaaSConfig.isSkipFunctionCall()) {
-                // when skipping the openFaaS function just pass on the original cloudEvent
-                kafkaMessageSender.safeSendCloudEvent(cloudEvent, originalMessageId);
             } else {
-                String functionResult = callFaasFunction(cloudEvent);
-                ArrayList<MicoCloudEventImpl<JsonNode>> events = parseFunctionResult(functionResult, cloudEvent);
+                List<MicoCloudEventImpl<JsonNode>> events = faasController.callFaasFunction(cloudEvent);
                 for (MicoCloudEventImpl<JsonNode> event : events) {
                     // individually wrap each cloud event in the results for sending
                     kafkaMessageSender.safeSendCloudEvent(event, originalMessageId);
@@ -98,48 +74,6 @@ public class MessageListener {
             }
         } catch (MicoCloudEventException e) {
             kafkaMessageSender.safeSendErrorMessage(e.getErrorEvent(), this.kafkaConfig.getInvalidMessageTopic(), originalMessageId);
-        }
-    }
-
-    /**
-     * Synchronously call the configured openFaaS function.
-     *
-     * @param cloudEvent the cloud event used as parameter for the function
-     * @return the result of the function call (in serialized form)
-     */
-    public String callFaasFunction(MicoCloudEventImpl<JsonNode> cloudEvent) throws MicoCloudEventException {
-        try {
-            URL functionUrl = openFaaSConfig.getFunctionUrl();
-            log.debug("Start request to function '{}'", functionUrl.toString());
-            String cloudEventSerialized = Json.encode(cloudEventManipulator.updateRouteHistoryWithFunctionCall(cloudEvent, openFaaSConfig.getFunctionName()));
-            log.debug("Serialized cloud event: {}", cloudEventSerialized);
-            String result = restTemplate.postForObject(functionUrl.toString(), cloudEventSerialized, String.class);
-            log.debug("Faas call resulted in: '{}'", result);
-            return result;
-        } catch (MalformedURLException e) {
-            throw new MicoCloudEventException("Failed to call faas-function. Caused by: " + e.getMessage(), cloudEvent);
-        } catch (IllegalStateException e) {
-            log.error("Failed to serialize CloudEvent '{}'.", cloudEvent);
-            throw new MicoCloudEventException("Failed to serialize CloudEvent while calling the faas-function.", cloudEvent);
-        } catch (HttpStatusCodeException e) {
-            log.error("A client error occurred with http status:{} . These exceptions are triggered if the  FaaS function does not return 200 OK as the status code", e.getStatusCode(), e);
-            throw new MicoCloudEventException(e.toString(), cloudEvent);
-        }
-    }
-
-    /**
-     * Parse the result of a faas function call.
-     *
-     * @param sourceCloudEvent only used for better error messages
-     * @return an ArrayList of cloud events
-     */
-    public ArrayList<MicoCloudEventImpl<JsonNode>> parseFunctionResult(String functionResult, MicoCloudEventImpl<JsonNode> sourceCloudEvent) throws MicoCloudEventException {
-        try {
-            return Json.decodeValue(functionResult, new TypeReference<ArrayList<MicoCloudEventImpl<JsonNode>>>() {
-            });
-        } catch (IllegalStateException e) {
-            log.error("Failed to parse JSON from response '{}'.", functionResult);
-            throw new MicoCloudEventException("Failed to parse JSON from response from the faas-function.", sourceCloudEvent);
         }
     }
